+++ lib/clixon/clixon_xml.h
@@ -273,11 +273,8 @@ cxobj    *xml_child_i(cxobj *xn, int i);
 cxobj    *xml_child_i_type(cxobj *xn, int i, enum cxobj_type type);
 cxobj    *xml_child_i_set(cxobj *xt, int i, cxobj *xc);
 int       xml_child_order(cxobj *xn, cxobj *xc);
-#ifndef XML_XCHILD_USE_ITER33
-int       xml_vector_decrement(cxobj *x, int nr);
-#endif
 cxobj    *xml_child_each(cxobj *xparent, cxobj *xprev,  enum cxobj_type type);
-cxobj    *xml_child_each_attr(cxobj *xparent, cxobj *xprev);
+cxobj    *xml_child_iter_attr(cxobj *xparent, int *inext);
 cxobj    *xml_child_iter(cxobj *xparent, int *inext, enum cxobj_type type);
 int       xml_child_insert_pos(cxobj *x, cxobj *xc, int pos);
 int       xml_childvec_set(cxobj *x, int len);