+++ lib/src/clixon_xml_nsctx.c
@@ -260,11 +260,13 @@ xml_nsctx_node1(cxobj *xn,
     char  *nm;  /* name */
     char  *val; /* value */
     cxobj *xp;  /* parent */
+    int    ix;
 
     /* xmlns:t="<ns1>" prefix:xmlns, name:t
      * xmlns="<ns2>"   prefix:NULL   name:xmlns
      */
-    while ((xa = xml_child_each_attr(xn, xa)) != NULL){
+    ix = 0;
+    while ((xa = xml_child_iter_attr(xn, &ix)) != NULL){
         pf = xml_prefix(xa);
         nm = xml_name(xa);
         if (pf == NULL){
@@ -691,11 +693,12 @@ xml2prefix(cxobj      *xn,
     char  *prefix = NULL;
     char  *xaprefix;
     int    ret;
+    int    ix;
 
     if (nscache_get_prefix(xn, namespace, &prefix) == 1) /* found */
         goto found;
-    xa = NULL;
-    while ((xa = xml_child_each_attr(xn, xa)) != NULL) {
+    ix = 0;
+    while ((xa = xml_child_iter_attr(xn, &ix)) != NULL) {
         /* xmlns=namespace */
         if (strcmp("xmlns", xml_name(xa)) == 0){
             if (strcmp(xml_value(xa), namespace) == 0){